[AP] Reorganize code for matmul_pass.
